Keep cast for norm

+++ pipeline_model.py
@@ -193,9 +193,8 @@ class OutputLayer(nn.Module):
         with torch.no_grad():
             log_vocab_size = math.log(logits.size(-1))
             # Compute the norms of the hidden states
-            hidden_state_norms = torch.norm(hidden_states, dim=-1)
+            hidden_state_norms = torch.norm(hidden_states.float(), dim=-1)
             hidden_state_norms = hidden_state_norms.view(-1)[flat_loss_mask]
-            hidden_state_norms = hidden_state_norms.float()
             # Compute the normalised entropy
             entropy = entropy_fn(flat_logits)[flat_loss_mask]
             # Compute normalised entropy so we can compare between models with different vocab sizes
